requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
python-dotenv>=0.19.0
asyncio>=3.4.3
psycopg2-binary>=2.9.1
transformers>=4.30.0
torch>=2.0.0
numpy==1.21.0
pandas>=1.3.0
sentence-transformers>=2.2.0
pytest == 8.3.4
httpx==0.28.1
common==0.1.2
dual==0.0.10
tight==0.1.0
data==0.4
prox==0.0.17
paddle==1.0.2.
paddlenlp==2.6.1
